backend/services/analytics_service.py
# app/services/analytics_service.py - Analytics calculations
from app.extensions import db
from app.models import SiteVisit, Booking
from sqlalchemy import func, extract
from datetime import datetime, timedelta

class AnalyticsService:
    @staticmethod
    def get_visit_stats():
        """Calculate visit statistics"""
        now = datetime.utcnow()
        today = now.date()
        
        # Calculate daily, weekly, monthly, yearly visits
        daily_visits = SiteVisit.query.filter(
            func.date(SiteVisit.timestamp) == today
        ).count()
        
        # ... rest of analytics logic
        
        return {
            'daily': daily_visits,
            'weekly': weekly_visits,
            'monthly': monthly_visits,
            'yearly': yearly_visits
        }
    
    @staticmethod
    def get_booking_stats():
        """Calculate booking statistics"""
        # Booking analytics logic
        pass
    
    @staticmethod
    def get_popular_destinations():
        """Get most popular destinations"""
        # Popular destinations logic
        pass