backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
bcrypt==4.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
gunicorn==21.2.0
redis==5.0.0
celery==5.3.4